build(pypi): Publish package to PyPI
Problem:
Per Blender Extension review, we were asked to publish this dependent
python client package to PyPI, as any Wheel used in the extension must
be published there.
https://extensions.blender.org/approval-queue/text2motion/#activity-5144

Solution:
Change the build process for this package such that it publish to PyPI
instead of to the Release Artifact in GitHub.

Following adjustments were made:
1. Building only the Wheel file doesn't requires any python dependencies
in the previous build. Now that we are building to publish to PyPI,
we actually needed to run `python3 -m build`, which requires `build`
along with several other build-tool dependencies. To accommodate this,
we added a Dockerfile with the necessary dependencies in
requirements.txt.
2. When we tested publishing to test.pypi, we notice the README.md was
not rendered at all. We ended up having to patch the setup.py using
block find and replace and swap out the `long_description` with the
README.md file content, as well as filling out other contents like
Author and our HomePage.
3. The README.md file generated by OpenAPI uses relative link, which
does not work in PyPI. Since the README.md is only found inside of our
build output, we added a note to redirect the user to download the
source distribution if they want to read the full README.md.
4. GitHub Action now publishes the package to PyPI instead of uploading
wheel to GitHub Release Artifact.
5. Now that we no longer want Releases Artifact on Github, added some
tag to link to PyPI package instead. The link will not work until we
fully publish the package to PyPI.

Documentations:
- https://packaging.python.org/en/latest/tutorials/packaging-projects/
- https://github.com/pypa/gh-action-pypi-publish

Risk:
This modifies the behavior of the build process from now and onwards.
However, it should not impact existing process as they are pinned on
the old version.

Testing Done:
1. Tested running the build locally using
`PACKAGE_VERSION=0.1.6 ./generate.sh`. Verified the build succeeded and
the patches works as expected.
2. Tested the new GHA by adding a manual dispatch. Verified the
publishing process works as expected.
https://test.pypi.org/project/text2motion-client-api/
3. Directly install the Build Distribution whl file from test.pypi into
local blender extension build and verified server request still works.
https://test.pypi.org/project/text2motion-client-api/#files
4. Tested installing the package from test.pypi and verified
installation succeeded.